chore: resolve merge comments
